face_auth_system/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
bcrypt
pydantic[email]
python-multipart
PyJWT
twilio
numpy
scipy
soundfile
torch
torchaudio
speechbrain
deepface
tf-keras
python-jose[cryptography]
